# Add faster raycast algorithm using DDA