--- a/include/mcl_3dl/point_cloud_random_samplers/point_cloud_sampler_with_normal.h
+++ b/include/mcl_3dl/point_cloud_random_samplers/point_cloud_sampler_with_normal.h
@@ -118,7 +118,7 @@ public:
     typename pcl::PointCloud<POINT_TYPE>::Ptr output(new pcl::PointCloud<POINT_TYPE>);
     output->header = pc->header;
 
-    if (pc->points.size() == 0)
+    if ((pc->points.size() == 0) || (num == 0))
     {
       return output;
     }