--- a/src/lidar_measurement_model_beam.cpp
+++ b/src/lidar_measurement_model_beam.cpp
@@ -29,6 +29,7 @@
 
 #include <algorithm>
 #include <cmath>
+#include <memory>
 #include <string>
 #include <vector>