--- a/src/lidar_measurement_model_beam.cpp
+++ b/src/lidar_measurement_model_beam.cpp
@@ -245,7 +245,7 @@ LidarMeasurementModelBeam::BeamStatus LidarMeasurementModelBeam::getBeamStatus(
     }
     else
     {
-      return BeamStatus::TOTAL_REFRECTION;
+      return BeamStatus::TOTAL_REFLECTION;
     }
   }
   return BeamStatus::LONG;