--- a/src/mcl_3dl.cpp
+++ b/src/mcl_3dl.cpp
@@ -135,7 +135,7 @@ protected:
     }
 
     pc_map_.reset(new pcl::PointCloud<PointType>);
-    const ros::Time map_stamp = (msg->header.stamp == ros::Time()) ? ros::Time::now() : msg->header.stamp;
+    const ros::Time map_stamp = msg->header.stamp.isValid() ? msg->header.stamp : ros::Time::now();
     pcl_conversions::toPCL(map_stamp, pc_map_->header.stamp);
     pc_map2_.reset();
     pc_update_.reset();